Add Carousel.spec
